Fixed language mistakes, update push to support Table and TableSchema, combine SQL command in to single method for BHoMObject and SQLIte_oM objects

+++ SQLite_Engine/Query/TableExists.cs
@@ -53,7 +53,7 @@ namespace BH.Engine.SQLite
                 using (var command = new SqliteCommand(checkSql, connection))
                 {
                     command.Parameters.AddWithValue("@TableName", tableName);
-                    int count = (int)command.ExecuteScalar();
+                    long count = (long)command.ExecuteScalar();
                     return count > 0;
                 }
             }
